PR comments refactors
